NIFI-13439 Addressing review comments
