+++ nifi-extension-bundles/nifi-sql-reporting-bundle/nifi-sql-reporting-tasks/src/test/java/org/apache/nifi/reporting/sql/TestQueryNiFiReportingTask.java
@@ -502,7 +502,7 @@ class TestQueryNiFiReportingTask {
         assertEquals(4, rows.size());
         // Validate the first row
         Map<String, Object> row = rows.get(0);
-        assertEquals(33, row.size());
+        assertEquals(26, row.size());
         assertEquals(1L, row.get("bytesRead"));
         // Validate the second row
         row = rows.get(1);
