+++ nifi-framework-bundle/nifi-framework/nifi-client-dto/src/main/java/org/apache/nifi/web/api/dto/status/ProcessGroupStatusSnapshotDTO.java
@@ -17,7 +17,7 @@
 package org.apache.nifi.web.api.dto.status;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import org.apache.nifi.controller.status.PerformanceMetrics;
+import org.apache.nifi.controller.status.ProcessingPerformanceStatus;
 import org.apache.nifi.web.api.entity.ConnectionStatusSnapshotEntity;
 import org.apache.nifi.web.api.entity.PortStatusSnapshotEntity;
 import org.apache.nifi.web.api.entity.ProcessGroupStatusSnapshotEntity;
@@ -82,7 +82,7 @@ public class ProcessGroupStatusSnapshotDTO implements Cloneable {
     private Integer terminatedThreadCount = 0;
     private Long processingNanos = 0L;
 
-    private PerformanceMetrics performanceMetrics;
+    private ProcessingPerformanceStatus processingPerformanceStatus;
 
     /**
      * The id for the process group.
@@ -523,12 +523,12 @@ public class ProcessGroupStatusSnapshotDTO implements Cloneable {
         this.processingNanos = processingNanos;
     }
 
-    public PerformanceMetrics getPerformanceMetrics() {
-        return performanceMetrics;
+    public ProcessingPerformanceStatus getProcessingPerformanceStatus() {
+        return processingPerformanceStatus;
     }
 
-    public void setPerformanceMetrics(PerformanceMetrics performanceMetrics) {
-        this.performanceMetrics = performanceMetrics;
+    public void setProcessingPerformanceStatus(ProcessingPerformanceStatus processingPerformanceStatus) {
+        this.processingPerformanceStatus = processingPerformanceStatus;
     }
 
     @Override
@@ -609,7 +609,7 @@ public class ProcessGroupStatusSnapshotDTO implements Cloneable {
             other.setProcessGroupStatusSnapshots(childGroups);
         }
 
-        other.setPerformanceMetrics(getPerformanceMetrics());
+        other.setProcessingPerformanceStatus(getProcessingPerformanceStatus());
 
         return other;
     }
