+++ nifi-framework-bundle/nifi-framework/nifi-framework-components/src/main/java/org/apache/nifi/reporting/AbstractEventAccess.java
@@ -40,7 +40,7 @@ import org.apache.nifi.controller.status.ConnectionStatus;
 import org.apache.nifi.controller.status.LoadBalanceStatus;
 import org.apache.nifi.controller.status.PortStatus;
 import org.apache.nifi.controller.status.ProcessGroupStatus;
-import org.apache.nifi.controller.status.PerformanceMetrics;
+import org.apache.nifi.controller.status.ProcessingPerformanceStatus;
 import org.apache.nifi.controller.status.ProcessorStatus;
 import org.apache.nifi.controller.status.RemoteProcessGroupStatus;
 import org.apache.nifi.controller.status.RunStatus;
@@ -163,7 +163,8 @@ public abstract class AbstractEventAccess implements EventAccess {
         int flowFilesTransferred = 0;
         long bytesTransferred = 0;
         long processingNanos = 0;
-        final PerformanceMetrics metrics = new PerformanceMetrics();
+        final ProcessingPerformanceStatus performanceStatus = new ProcessingPerformanceStatus();
+        performanceStatus.setIdentifier(group.getIdentifier());
 
         final boolean populateChildStatuses = currentDepth <= recursiveStatusDepth;
 
@@ -195,7 +196,17 @@ public abstract class AbstractEventAccess implements EventAccess {
 
             processingNanos += procStat.getProcessingNanos();
 
-            PerformanceMetrics.merge(metrics, procStat.getPerformanceMetrics());
+            final ProcessingPerformanceStatus processorPerformanceStatus = procStat.getProcessingPerformanceStatus();
+
+            if (processorPerformanceStatus != null) {
+                performanceStatus.setCpuTime(performanceStatus.getCpuTime() + processorPerformanceStatus.getCpuTime());
+                performanceStatus.setReadTime(performanceStatus.getReadTime() + processorPerformanceStatus.getReadTime());
+                performanceStatus.setWriteTime(performanceStatus.getWriteTime() + processorPerformanceStatus.getWriteTime());
+                performanceStatus.setCommitTime(performanceStatus.getCommitTime() + processorPerformanceStatus.getCommitTime());
+                performanceStatus.setGcTime(performanceStatus.getGcTime() + processorPerformanceStatus.getGcTime());
+                performanceStatus.setBytesRead(performanceStatus.getBytesRead() + processorPerformanceStatus.getBytesRead());
+                performanceStatus.setBytesWritten(performanceStatus.getBytesWritten() + processorPerformanceStatus.getBytesWritten());
+            }
         }
 
         // set status for local child groups
@@ -230,7 +241,18 @@ public abstract class AbstractEventAccess implements EventAccess {
             bytesTransferred += childGroupStatus.getBytesTransferred();
 
             processingNanos += childGroupStatus.getProcessingNanos();
-            PerformanceMetrics.merge(metrics, childGroupStatus.getPerformanceMetrics());
+
+            final ProcessingPerformanceStatus childGroupPerformanceStatus = childGroupStatus.getProcessingPerformanceStatus();
+
+            if (childGroupPerformanceStatus != null) {
+                performanceStatus.setCpuTime(performanceStatus.getCpuTime() + childGroupPerformanceStatus.getCpuTime());
+                performanceStatus.setReadTime(performanceStatus.getReadTime() + childGroupPerformanceStatus.getReadTime());
+                performanceStatus.setWriteTime(performanceStatus.getWriteTime() + childGroupPerformanceStatus.getWriteTime());
+                performanceStatus.setCommitTime(performanceStatus.getCommitTime() + childGroupPerformanceStatus.getCommitTime());
+                performanceStatus.setGcTime(performanceStatus.getGcTime() + childGroupPerformanceStatus.getGcTime());
+                performanceStatus.setBytesRead(performanceStatus.getBytesRead() + childGroupPerformanceStatus.getBytesRead());
+                performanceStatus.setBytesWritten(performanceStatus.getBytesWritten() + childGroupPerformanceStatus.getBytesWritten());
+            }
         }
 
         // set status for remote child groups
@@ -518,7 +540,7 @@ public abstract class AbstractEventAccess implements EventAccess {
         status.setFlowFilesTransferred(flowFilesTransferred);
         status.setBytesTransferred(bytesTransferred);
         status.setProcessingNanos(processingNanos);
-        status.setPerformanceMetrics(metrics);
+        status.setProcessingPerformanceStatus(performanceStatus);
 
         final VersionControlInformation vci = group.getVersionControlInformation();
         if (vci != null) {
@@ -667,8 +689,8 @@ public abstract class AbstractEventAccess implements EventAccess {
                 status.setCounters(flowFileEvent.getCounters());
             }
 
-            final PerformanceMetrics metrics = PerformanceMetricsUtil.getPerformanceMetrics(flowFileEventRepository.reportAggregateEvent(), flowFileEvent, procNode);
-            status.setPerformanceMetrics(metrics);
+            final ProcessingPerformanceStatus performanceStatus = PerformanceMetricsUtil.getPerformanceMetrics(flowFileEvent, procNode);
+            status.setProcessingPerformanceStatus(performanceStatus);
         }
 
         // Determine the run status and get any validation error... only validating while STOPPED
