+++ nifi-framework-bundle/nifi-framework/nifi-framework-components/src/main/java/org/apache/nifi/reporting/PerformanceMetricsUtil.java
@@ -18,45 +18,22 @@ package org.apache.nifi.reporting;
 
 import org.apache.nifi.controller.ProcessorNode;
 import org.apache.nifi.controller.repository.FlowFileEvent;
-import org.apache.nifi.controller.status.PerformanceMetrics;
+import org.apache.nifi.controller.status.ProcessingPerformanceStatus;
 
 public class PerformanceMetricsUtil {
 
-    public static PerformanceMetrics getPerformanceMetrics(final FlowFileEvent aggregatedEvent, final FlowFileEvent fileEvent, final ProcessorNode processorNode) {
-        long totalCpuNanos = aggregatedEvent.getCpuNanoseconds();
-        long totalReadNanos = aggregatedEvent.getContentReadNanoseconds();
-        long totalWriteNanos = aggregatedEvent.getContentWriteNanoseconds();
-        long totalSessionCommitNanos = aggregatedEvent.getSessionCommitNanoseconds();
-        long totalBytesRead = aggregatedEvent.getBytesRead();
-        long totalBytesWritten = aggregatedEvent.getBytesWritten();
-        long totalGcNanos = aggregatedEvent.getGargeCollectionMillis();
-
-        final PerformanceMetrics newMetrics = new PerformanceMetrics();
+    public static ProcessingPerformanceStatus getPerformanceMetrics(final FlowFileEvent fileEvent, final ProcessorNode processorNode) {
+        final ProcessingPerformanceStatus newMetrics = new ProcessingPerformanceStatus();
 
         newMetrics.setIdentifier(processorNode.getProcessGroup().getIdentifier());
         newMetrics.setCpuTime(fileEvent.getCpuNanoseconds());
-        newMetrics.setCpuTimePercentage(nanosToPercent(fileEvent.getCpuNanoseconds(), totalCpuNanos));
         newMetrics.setReadTime(fileEvent.getContentReadNanoseconds());
-        newMetrics.setReadTimePercentage(nanosToPercent(fileEvent.getContentReadNanoseconds(), totalReadNanos));
         newMetrics.setWriteTime(fileEvent.getContentWriteNanoseconds());
-        newMetrics.setWriteTimePercentage(nanosToPercent(fileEvent.getContentWriteNanoseconds(), totalWriteNanos));
         newMetrics.setCommitTime(fileEvent.getSessionCommitNanoseconds());
-        newMetrics.setCommitTimePercentage(nanosToPercent(fileEvent.getSessionCommitNanoseconds(), totalSessionCommitNanos));
         newMetrics.setGcTime(fileEvent.getGargeCollectionMillis());
-        newMetrics.setGcTimePercentage(nanosToPercent(fileEvent.getGargeCollectionMillis(), totalGcNanos));
         newMetrics.setBytesRead(fileEvent.getBytesRead());
-        newMetrics.setBytesReadPercentage(nanosToPercent(fileEvent.getBytesRead(), totalBytesRead));
         newMetrics.setBytesWritten(fileEvent.getBytesWritten());
-        newMetrics.setBytesWrittenPercentage(nanosToPercent(fileEvent.getBytesWritten(), totalBytesWritten));
 
         return newMetrics;
-    }
-
-    private static long nanosToPercent(final long a, final long b) {
-        if (b < 1) {
-            return b;
-        } else {
-            return Math.min(100L, a * 100 / b);
-        }
     }
 }
